# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  common_models.FilterOperator:
    enum:
    - IS
    - IS_NOT
    - IS_ONE_OF_TERMS
    - IS_ONE_OF_TERMS_OR
    - IS_LESS_THAN_OR_EQUALS
    - IS_GREATER_THAN
    - EXIST
    - DOES_NOT_EXIST
    - NOT_CONTAINS
    - CONTAIN
    - DOES_NOT_CONTAIN
    - CONTAIN_ONE_OF
    - DOES_NOT_CONTAIN_ONE_OF
    - IS_ONE_OF
    - IS_NOT_ONE_OF
    - IS_BETWEEN
    - IS_NOT_BETWEEN
    - IS_IN_FIELDS
    - IS_NOT_IN_FIELDS
    - ENDS_WITH
    - NOT_ENDS_WITH
    - START_WITH
    - NOT_START_WITH
    type: string
    x-enum-varnames:
    - OpEquals
    - OpNotEquals
    - OpIn
    - OpInOr
    - OpLessOrEq
    - OpGreater
    - OpExists
    - OpNotExists
    - OpNotContains
    - OpContain
    - OpDoesNotContain
    - OpContainOneOf
    - OpDoesNotContainOneOf
    - OpIsOneOf
    - OpIsNotOneOf
    - OpIsBetween
    - OpIsNotBetween
    - OpIsInFields
    - OpIsNotInFields
    - OpEndsWith
    - OpNotEndsWith
    - OpStartWith
    - OpNotStartWith
  common_models.FilterType:
    properties:
      field:
        type: string
      operator:
        $ref: '#/definitions/common_models.FilterOperator'
      value: {}
    type: object
...
  dto.ChartViewRequest:
    properties:
      field:
        type: string
      fieldDataType:
        type: string
      filters:
        items:
          $ref: '#/definitions/common_models.FilterType'
        type: array
      indexPattern:
        type: string
      interval:
        type: string
      top:
        type: integer
    type: object
  dto.ChartViewResponse:
    properties:
      categories:
        items:
          type: string
        type: array
      values:
        items:
          type: integer
        type: array
    type: object
...
  dto.CreateTenantConfigRequest:
    properties:
      assetAvailability:
        type: integer
      assetConfidentiality:
        type: integer
      assetHostnameList:
        items:
          type: string
        type: array
      assetIntegrity:
        type: integer
      assetIpList:
        items:
          type: string
        type: array
      assetName:
        type: string
    required:
    - assetName
    type: object
...
  dto.TenantConfigResponse:
    properties:
      assetAvailability:
        type: integer
      assetConfidentiality:
        type: integer
      assetHostnameList:
        items:
          type: string
        type: array
      assetIntegrity:
        type: integer
      assetIpList:
        items:
          type: string
        type: array
      assetName:
        type: string
    type: object
...
  dto.TopValue:
    properties:
      count:
        type: integer
      percent:
        type: number
      value:
        type: string
    type: object
  dto.TopValuesResponse:
    properties:
      top:
        items:
          $ref: '#/definitions/dto.TopValue'
        type: array
      total:
        type: integer
    type: object
...
  dto.UpdateTenantConfigRequest:
    properties:
      assetAvailability:
        type: integer
      assetConfidentiality:
        type: integer
      assetHostnameList:
        items:
          type: string
        type: array
      assetIntegrity:
        type: integer
      assetIpList:
        items:
          type: string
        type: array
      assetName:
        type: string
    required:
    - assetName
    type: object
...
info:
  contact: {}
  description: UTMStack backend API — manages IAM, modules, flows, incidents, compliance
    and audit data.
  title: UTMStack Backend Service
  version: "1.0"
paths:
...
  /eventprocessing/tenant-config:
    get:
      parameters:
      - description: Page (0-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Partial match on assetName
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: Total records
              type: string
          schema:
            items:
              $ref: '#/definitions/dto.TenantConfigResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List assets (tenant configs)
      tags:
      - Event Processing
    post:
      consumes:
      - application/json
      parameters:
      - description: assetName + CIA + lists
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTenantConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TenantConfigResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create asset (tenant config)
      tags:
      - Event Processing
    put:
      consumes:
      - application/json
      parameters:
      - description: assetName + CIA + lists
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTenantConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TenantConfigResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update asset (tenant config)
      tags:
      - Event Processing
  /eventprocessing/tenant-config/{id}:
    delete:
      parameters:
      - description: Asset name
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete asset (tenant config)
      tags:
      - Event Processing
    get:
      parameters:
      - description: Asset name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TenantConfigResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get asset by name
      tags:
      - Event Processing
...
  /log-analyzer/chart-view:
    post:
      consumes:
      - application/json
      description: Returns category/value arrays for the explorer chart — a date_histogram
        when `interval` is set, otherwise a terms aggregation over `field`.
      parameters:
      - description: Chart view request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ChartViewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ChartViewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Chart/timeline aggregation
      tags:
      - Log Analyzer
...
  /log-analyzer/top-x-values/{indexPattern}/{field}/{top}:
    post:
      consumes:
      - application/json
      description: Returns the most frequent values of a field over an index (terms
        aggregation) plus the total doc count. Body is an optional filter list.
      parameters:
      - description: Index pattern
        in: path
        name: indexPattern
        required: true
        type: string
      - description: Field to aggregate
        in: path
        name: field
        required: true
        type: string
      - description: Number of top values
        in: path
        name: top
        required: true
        type: integer
      - description: Filters to apply
        in: body
        name: filters
        schema:
          items:
            $ref: '#/definitions/common_models.FilterType'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TopValuesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Top-N values of a field
      tags:
      - Log Analyzer
...
  /sso/saml/{name}/acs:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Receives the IdP SAMLResponse, validates it, maps the NameID to
        a local user and redirects to the app with a session token.
      parameters:
      - description: Identity provider name
        in: path
        name: name
        required: true
        type: string
      - description: Base64 SAML response
        in: formData
        name: SAMLResponse
        required: true
        type: string
      responses:
        "302":
          description: Redirect to the app with ?error=saml2
      summary: SAML assertion consumer service
      tags:
      - SSO
  /sso/saml/{name}/login:
    get:
      description: Redirects the browser to the IdP with a signed AuthnRequest for
        the named provider.
      parameters:
      - description: Identity provider name
        in: path
        name: name
        required: true
        type: string
      responses:
        "302":
          description: Redirect to the app with ?error=saml2
      summary: Start SAML SSO login
      tags:
      - SSO
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"